@@ -22,16 +22,16 @@
     ],
     "autoload": {
         "psr-4": {
-            "Comodojo\\Cookies\\": "src"
+            "Comodojo\\Cookies\\": "src/Comodojo/Cookies"
         }
     },
     "require": {
-        "php": ">=5.3.0",
-        "phpseclib/phpseclib": "0.3.*",
-        "comodojo/exceptions": "@stable"
+        "php": ">=5.4",
+        "phpseclib/phpseclib": "^2.0",
+        "comodojo/exceptions": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0",
-        "scrutinizer/ocular": "~1.1"
+        "phpunit/phpunit": "^4.0|^5.0",
+        "scrutinizer/ocular": "^1.0"
     }
 }
