@@ -1,18 +1,16 @@
 <?php namespace Comodojo\Cookies;
 
-use \Comodojo\Cookies\CookieInterface\CookieInterface;
 use \Comodojo\Exception\CookieException;
-use \Comodojo\Cookies\CookieBase;
 
 /**
  * Plain cookie
- * 
+ *
  * @package     Comodojo Spare Parts
  * @author      Marco Giovinazzi <marco.giovinazzi@comodojo.org>
  * @license     MIT
  *
  * LICENSE:
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -22,40 +20,7 @@ use \Comodojo\Cookies\CookieBase;
  * THE SOFTWARE.
  */
 
-class Cookie extends CookieBase implements CookieInterface {
-
-    /**
-     * Cookie constructor
-     *
-     * Setup cookie name
-     *
-     * @param  string   $name
-     *
-     * @throws \Comodojo\Exception\CookieException
-     */
-    public function __construct($name) {
-
-        try {
-
-            $this->setName($name);
-            
-        } catch (CookieException $ce) {
-            
-            throw $ce;
-
-        }
-
-        if ( defined("COMODOJO_COOKIE_MAX_SIZE") ) {
-
-            $this->max_cookie_size = filter_var(COMODOJO_COOKIE_MAX_SIZE, FILTER_VALIDATE_INT, array(
-                'options' => array(
-                    'default' => 4000
-                )
-            ));
-
-        }
-
-    }
+class Cookie extends AbstractCookie {
 
     /**
      * Set cookie content
@@ -69,11 +34,11 @@ class Cookie extends CookieBase implements CookieInterface {
      */
     public function setValue($value, $serialize = true) {
 
-        if ( !is_scalar($value) AND $serialize === false ) throw new CookieException("Cannot set non-scalar value without serialization");
+        if ( !is_scalar($value) && $serialize === false ) throw new CookieException("Cannot set non-scalar value without serialization");
 
         $cookie_value = $serialize === true ? serialize($value) : $value;
 
-        if ( strlen($cookie_value) > $this->max_cookie_size ) throw new CookieException("Cookie size larger than 4KB");
+        if ( strlen($cookie_value) > $this->max_cookie_size ) throw new CookieException("Cookie size larger than ".$this->max_cookie_size." bytes");
 
         $this->value = $cookie_value;
 
@@ -98,23 +63,25 @@ class Cookie extends CookieBase implements CookieInterface {
      * Static method to create a cookie quickly
      *
      * @param   string   $name  The cookie name
-     * 
+     *
      * @param   array    $properties    Array of properties cookie should have
      *
      * @return  \Comodojo\Cookies\Cookie
      *
      * @throws  \Comodojo\Exception\CookieException
      */
-    public static function create($name, $properties = array(), $serialize = true) {
+    public static function create($name, $properties = [], $serialize = true) {
 
         try {
 
-            $cookie = new Cookie($name);
+            $class = get_called_class();
+
+            $cookie = new $class($name);
 
             self::cookieProperties($cookie, $properties, $serialize);
 
         } catch (CookieException $ce) {
-            
+
             throw $ce;
 
         }
@@ -136,12 +103,14 @@ class Cookie extends CookieBase implements CookieInterface {
 
         try {
 
-            $cookie = new Cookie($name);
+            $class = get_called_class();
+
+            $cookie = new $class($name);
 
             $return = $cookie->load();
 
         } catch (CookieException $ce) {
-            
+
             throw $ce;
 
         }
